Work-around for NLopt solver getting stuck (#12)
This happens very rarely, and appears to be related to numerical
issues (e.g., time slices that have no coalescences in them, etc),
but it is annoying when it does:
the solver never improves on the minimum objective function values,
and I suspect it is just ping-ponging between values internally.

Now we kill the solver if it runs for 2,000 iterations without
improving upon the minimum value.