Extend population splits to support admixture (#14)
* Make mrpast input model YAML more user-friendly

The vector/matrix representation was convenient (kind of...) for
working with in the solver, but not particularly easy to understand
or modify for users.

Instead, use a representation where we just declare the relationships,
so you don't have to put '0' for "no parameter", etc.

* Demes conversion uses the true value to generate the bounds,
   so we avoid violating bounds.
* The popname is used in demes conversion, and is generally usable
   in the mrpast models
* Added more tests
* Support for fixed parameters
* Demes -> mrpast conversion now handles admixture properly

Given an admixture from N ancestral demes, there are (N-1) free
parameters and the remaining Nth is determined by the others.
The solver now supports this, with the "one-minus" parameter type.

* Add example based on AmericanAdmixture_4B18
* I cleaned up the solver parameter handling a bit, but it still
could use some refactoring.
* Static linking makes the Python packaging much simpler
* Add note in the README about the model format change