Don't allow admixture values to become negative
Some admixture parameters are determined completely by the sum of
other admixture parameters (1.0 - sum, actually). Depending on the
bounds chosen by the user, this can lead to negative admixture
proportions. Prevent this from happening, and always zero out such
parameters.

It is still _possible_ that the sum of all admixture proportions
could exceed 1, but in practice for a reasonable model this does
not happen for the MLE, just for random initialization values.