Tighten epoch time bounds on non-synthetic models
For simulated ARGs, we can accurately infer most epoch times even
with huge (unnecessarily so) bounds. For inferred ARGs, it varies
by model, but especially if we treat certain populations as ghosts
the epoch times can be less accurate. For most of the events marked
by these epochs, we can rely on external knowledge from the model
to determine a rough idea of when it occurred, thus we don't need
bounds that are thousands of generations in size.

I shrink most of these bounds to +/- a few hundred generations on
either side. We have observed that this helps guide the solver to
solutions where most epoch times are _not_ reaching the bounds, but
producing overall better solutions (for all parameters).