Add flag to remap populations based on order
This is necessary when doing ghost populations in inferred ARGs.
E.g., the inferred ARG might have populations 0...3 that map to
certain populations in your model, but your model has 5 populations
and you need to tell the solver which ARG pops are for which
model pops. You will get a table of the mapping upon completion,
for sanity checking.